Improve libparsec_platform_async, unify native&web test
